Development: Extract PDF merge utility
